semantic_inference_python/requirements.txt
click==8.1.7
distinctipy==1.3.4
einops==0.8.0
imageio==2.35.1
onnx==1.17.0
pandas==2.0.3
pyyaml==6.0.2
seaborn==0.13.2
torch==2.4.1
torchvision==0.19.1
ultralytics==8.3.150
clip @ git+https://github.com/openai/CLIP.git
open_clip_torch==2.29.0
numpy<2
supervision==0.18.0
transformers==4.46.3
timm==1.0.11
accelerate==1.0.1
sentencepiece==0.2.0
attrdict==2.0.1
rospkg==1.5.1
openai==1.97.0
